+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * add `serverexception` event on `Netzke.directProvider`, subscribe to it to handle server exceptions
   * endpoint calls will now pass Ext.direct.Exception to the provided callback function in case of server exception
   * a component will clean-up the loading mask when a server exception occurs during dynamic component loading
+  * `netzkeFeedback` now shows multiple messages in a single slide banner
 
 # v0.8.3 - 2013-03-22
 * support Rails 3.2.13